Sustainable Shopping Assistant/requirements.txt
streamlit
streamlit-chat
openai
python-dotenv
pyttsx3
SpeechRecognition